pyproject.toml
[tool.poetry]
name = "schedulebot"
version = "0.1.0"
description = ""
authors = ["DTraitor <sehhd32@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-telegram-bot = "^20.5"
pytz = "^2023.3"
tabula-py = "^2.7.0"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
